update: rename Init to Splash
